Prashanth changes to introudce Junit as part of ENC developement. In high level all the required classes are loaded with the help of /META-INF/*test.xml and the caches are loaded using /META-INF/*ehcache.xml.
EncCatalog service extends the exisiting blc catalog service and only
introuduces test method and statement to check if unit testing is
working. This class creates a design and material on the run and tests
the same. Please make sure your db schema has got enc_design and
enc_material tables with just String 'TEST' coloumn alone to test the
junit completely. also the class queries the existing db for design and
material and prints their name in the console.